research_agent/requirements.txt
openai
anthropic
google-generativeai
tavily-python
python-dotenv
streamlit
gTTS
pydantic
tenacity
pydantic-settings
elevenlabs
fpdf
